requirements.txt
Django==3.2.6
whitenoise==5.3.0
gunicorn==20.1.0
dj-database-url==0.5.0
Pillow==7.0.0
django-map-widgets==0.3.2
django-google-maps==0.12.4
django-location==2.1.1
geocoder==1.38.1
social-auth-app-django==4.0.0
social-auth-core==4.1.0
haversine==2.3.1
xhtml2pdf==0.2.5
boto3==1.21.0
django-storages==1.12.3